fix return type.
Mon Feb 17 15:20:10 2003  Tim Janik  <timj@gtk.org>

        * sfimemory.[hc] (sfi_alloc_upper_power2): fix return type.

svn path=/trunk/; revision=896

+++ sfi/sfimemory.c
@@ -33,7 +33,7 @@ static gulong       memory_allocated = 0;
 
 
 /* --- functions --- */
-const guint
+const gulong
 sfi_alloc_upper_power2 (const gulong number)
 {
   return number ? 1 << g_bit_storage (number - 1) : 0;
